DPR-422 move helper code into test specific package
