+++ src/test/java/uk/gov/justice/digital/test/SparkTestHelpers.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.service;
+package uk.gov.justice.digital.test;
 
 import lombok.val;
 import org.apache.commons.io.FileUtils;
@@ -10,6 +10,7 @@ import org.apache.spark.sql.types.DataTypes;
 import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import uk.gov.justice.digital.domain.model.TableIdentifier;
+import uk.gov.justice.digital.service.DataStorageService;
 
 import java.io.IOException;
 import java.nio.file.Path;
